Move memory access analysis to polygeist
