+++ .github/workflows/sycl_linux_build_and_test.yml
@@ -111,6 +111,7 @@ jobs:
       run: cmake --build $GITHUB_WORKSPACE/build --target deploy-sycl-toolchain
     # TODO allow to optionally disable in-tree checks
     - name: check-mlir-sycl
+      shell: bash
       if: ${{ always() && !cancelled() && steps.build.outcome == 'success' }}
       run: |
         cmake --build $GITHUB_WORKSPACE/build --target check-mlir-sycl
@@ -123,9 +124,16 @@ jobs:
         fi
         cmake --build $GITHUB_WORKSPACE/build --target check-cgeist
     - name: check-polygeist-opt
+      shell: bash    
       if: ${{ always() && !cancelled() && steps.build.outcome == 'success' }}
       run: |
         cmake --build $GITHUB_WORKSPACE/build --target check-polygeist-opt
+    - name: MLIRMemoryAccessAnalysisTests
+      shell: bash    
+      if: ${{ always() && !cancelled() && steps.build.outcome == 'success' }}
+      run: |
+        cmake --build $GITHUB_WORKSPACE/build --target MLIRMemoryAccessAnalysisTests
+         ./tools/polygeist/unittests/Analysis/MemoryAccessAnalysis/MLIRMemoryAccessAnalysisTests
     - name: Install
       # TODO replace utility installation with a single CMake target
       run: |
