+++ mlir-sycl/CMakeLists.txt
@@ -75,13 +75,6 @@ add_subdirectory(lib)
 add_subdirectory(test)
 add_subdirectory(tools)
 
-add_custom_target(MLIRSYCLUnitTests)
-if (EXISTS ${LLVM_THIRD_PARTY_DIR}/unittest/googletest/include/gtest/gtest.h)
-  add_subdirectory(unittests)
-else()
-  message(WARNING "gtest not found, unittests will not be available")
-endif()
-
 install(DIRECTORY include/
   DESTINATION include
   COMPONENT sycl-mlir-headers
