+++ polygeist/CMakeLists.txt
@@ -100,3 +100,11 @@ add_subdirectory(include)
 add_subdirectory(lib)
 add_subdirectory(tools)
 add_subdirectory(test)
+
+add_custom_target(MLIRPolygeistUnitTests)
+if (EXISTS ${LLVM_THIRD_PARTY_DIR}/unittest/googletest/include/gtest/gtest.h)
+  add_subdirectory(unittests)
+else()
+  message(WARNING "gtest not found, unittests will not be available")
+endif()
+
